move app directory
