--- src/app/controllers/TransactionController.ts
@@ -1,8 +1,0 @@
-import * as DB from '../../../db';
-import * as Types from '../../lib/types';
-
-export class TransactionController {
-
-  
-
-}
